--- src/lib/firebase.auth-email.util.js
@@ -1,6 +1,0 @@
-import { getAuth, createUserWithEmailAndPassword, sendEmailVerification } from 'https://www.gstatic.com/firebasejs/9.6.8/firebase-auth.js';
-
-export {
-    getAuth,
-    createUserWithEmailAndPassword, sendEmailVerification
-}
