--- src/lib/firebase.auth-util.js
@@ -1,7 +1,0 @@
-import {
-  getAuth,
-  signInWithPopup,
-  GoogleAuthProvider,
-} from 'https://www.gstatic.com/firebasejs/9.6.8/firebase-auth.js';
-
-export { getAuth, signInWithPopup, GoogleAuthProvider };
