+++ src/lib/firebase.js
@@ -1,11 +1,16 @@
 // Import the functions you need from the SDKs you need
-import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.6.8/firebase-app.js';
-import { getAnalytics } from 'https://www.gstatic.com/firebasejs/9.6.8/firebase-analytics.js';
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
 
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
+import {
+  initializeApp,
+  getAnalytics,
+  getAuth,
+  createUserWithEmailAndPassword,
+  sendEmailVerification,
+  signInWithPopup,
+  GoogleAuthProvider,
+  signInWithEmailAndPassword,
+} from './firebase.util.js';
+
 const firebaseConfig = {
   apiKey: 'AIzaSyDIRrQ_iwja7mfLbiFMCFyAzdhDcgcsoRI',
   authDomain: 'paws-sn.firebaseapp.com',
@@ -21,4 +26,78 @@ const firebaseConfig = {
 export const iniciarFirebase = () => {
   const app = initializeApp(firebaseConfig);
   const analytics = getAnalytics(app);
+};
+
+// registrarse con cualquier correo
+
+export const signInEmail = (email, password) => {
+  const auth = getAuth();
+  createUserWithEmailAndPassword(auth, email, password)
+    .then((userCredential) => {
+      // Signed in
+      const user = userCredential.user;
+      console.log(user);
+
+      sendEmailVerification(auth.currentUser).then(() => {
+        alert('Hemos enviado a tu correo electrónico el enlace de confirmación');
+      });
+      window.location.hash = 'login';
+    })
+
+    .catch((error) => {
+      const errorCode = error.code;
+      alert('Correo ya registrado, por favor intente con otro correo.');
+      const errorMessage = error.message;
+
+      // ..
+    });
+};
+
+// registrarse o iniciar sesión con google
+
+export const signInGoogle = () => {
+  const provider = new GoogleAuthProvider();
+
+  const auth = getAuth();
+  signInWithPopup(auth, provider)
+    .then((result) => {
+      const credential = GoogleAuthProvider.credentialFromResult(result);
+      const token = credential.accessToken;
+      const user = result.user;
+      sessionStorage.setItem('token', token);
+      sessionStorage.setItem('user', JSON.stringify(user));
+
+      window.location.hash = 'post';
+      // ...
+    })
+    .catch((error) => {
+      const errorCode = error.code;
+      const errorMessage = error.message;
+      const email = error.email;
+      const credential = GoogleAuthProvider.credentialFromError(error);
+    });
+};
+// Iniciar sesión
+
+export const logInEmail = (email, password) => {
+  const auth = getAuth();
+  signInWithEmailAndPassword(auth, email, password)
+    .then((userCredential) => {
+      // Signed in
+      const user = userCredential.user;
+      // ...
+      // window.location.hash = 'post';
+
+      if (user.emailVerified) {
+        sessionStorage.setItem('user', JSON.stringify(user));
+        window.location.hash = 'post';
+      } else {
+        alert('Para iniciar sesión debes confirmar el link que enviamos a tu correo electrónico');
+      }
+    })
+    .catch((error) => {
+      const errorCode = error.code;
+      alert('Usario y/o contraseña inválido');
+      const errorMessage = error.message;
+    });
 };
